Re-organize the file structure: organize file by type
