+++ app/src/main/java/com/jmstudios/redmoon/Intro.kt
@@ -10,9 +10,6 @@ import androidx.fragment.app.Fragment
 import com.github.paolorotolo.appintro.AppIntro2
 import com.github.paolorotolo.appintro.AppIntroFragment
 
-import com.jmstudios.redmoon.R
-import com.jmstudios.redmoon.model.Config
-
 class Intro : AppIntro2() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
