+++ app/src/main/java/com/jmstudios/redmoon/Util.kt
@@ -29,12 +29,9 @@ import androidx.preference.Preference
 import androidx.core.content.ContextCompat
 import androidx.preference.PreferenceFragmentCompat
 
-import com.jmstudios.redmoon.RedMoonApplication
 import com.jmstudios.redmoon.helper.EventBus
 import com.jmstudios.redmoon.helper.KLog
 import com.jmstudios.redmoon.helper.KLogging
-import com.jmstudios.redmoon.model.Profile
-import com.jmstudios.redmoon.model.Config
 
 import java.util.Calendar
