+++ app/src/main/java/com/jmstudios/redmoon/activity/MainActivity.kt
@@ -36,9 +36,7 @@ import com.jmstudios.redmoon.fragment.FilterFragment
 import com.jmstudios.redmoon.helper.EventBus
 import com.jmstudios.redmoon.helper.Logger
 import com.jmstudios.redmoon.helper.Permission
-import com.jmstudios.redmoon.model.Config
-import com.jmstudios.redmoon.model.ProfilesModel
-import com.jmstudios.redmoon.util.*
+import com.jmstudios.redmoon.ProfilesModel
 
 import org.greenrobot.eventbus.Subscribe
