+++ app/src/main/java/com/jmstudios/redmoon/preference/ProfileSelectorPreference.kt
@@ -24,10 +24,10 @@ import com.jmstudios.redmoon.activeProfile
 import com.jmstudios.redmoon.getString
 
 import com.jmstudios.redmoon.helper.Logger
-import com.jmstudios.redmoon.model.Profile
-import com.jmstudios.redmoon.model.Config
-import com.jmstudios.redmoon.model.ProfilesModel
-import com.jmstudios.redmoon.model.ProfilesModel.isSaved
+import com.jmstudios.redmoon.Profile
+import com.jmstudios.redmoon.Config
+import com.jmstudios.redmoon.ProfilesModel
+import com.jmstudios.redmoon.ProfilesModel.isSaved
 import com.jmstudios.redmoon.profilesUpdated
 
 import org.greenrobot.eventbus.Subscribe
